Move plugins into schema/
