+++ src/schema/index.ts
@@ -2,7 +2,7 @@ import { makeSchema } from "nexus";
 import { nexusPrisma } from "nexus-plugin-prisma";
 import { join } from "path";
 
-import { argValidation } from "../plugins";
+import { argValidation } from "./plugins";
 import * as types from "./types";
 
 export const schema = makeSchema({
